+++ setting/report_emp.php
@@ -20,9 +20,17 @@
                 <div class="row">
                     <div class="col-12">
                         <div class="card">
-                            <div class="card-header">
-                                <button type="button" class="btn btn-primary waves-effect" data-toggle="modal" data-target="#add-Modal"> Add New Employee <i class="fa fa-plus"></i> </button>
-                            </div>
+                            <?php
+                            if (!isset($_SESSION['super_id'])) {
+                            ?>
+                                <div class="card-header">
+                                    <button type="button" class="btn btn-primary waves-effect" data-toggle="modal" data-target="#add-Modal"> Add New Employee <i class="fa fa-plus"></i> </button>
+                                </div>
+                            <?php
+                            }
+
+                            ?>
+
                             <?php
                             if (isset($_SESSION['success_message'])) {
                                 $message = $_SESSION['success_message'];
@@ -77,8 +85,16 @@
                                                     <td> <a href="main.php?dir=main_menu&page=emp_orders&emp_id=<?php echo $emp['id']; ?>" class="btn btn-success btn-sm"> View Orders </a> </td>
                                                     <td> <a href="main.php?dir=statistics&page=report&emp_id=<?php echo $emp['id']; ?>" class="btn btn-primary btn-sm"> Statistics </a> </td>
                                                     <td>
-                                                        <button type="button" class="btn btn-success btn-sm waves-effect" data-toggle="modal" data-target="#edit-Modal_<?php echo $emp['id']; ?>"> Edit <i class='fa fa-pen'></i> </button>
-                                                        <a href="main.php?dir=setting&page=delete_emp&emp_id=<?php echo $emp['id']; ?>" class="confirm btn btn-danger btn-sm"> Delete <i class='fa fa-trash'></i> </a>
+                                                        <?php
+                                                        if (!isset($_SESSION['super_id'])) {
+                                                        ?>
+                                                            <button type="button" class="btn btn-success btn-sm waves-effect" data-toggle="modal" data-target="#edit-Modal_<?php echo $emp['id']; ?>"> Edit <i class='fa fa-pen'></i> </button>
+                                                            <a href="main.php?dir=setting&page=delete_emp&emp_id=<?php echo $emp['id']; ?>" class="confirm btn btn-danger btn-sm"> Delete <i class='fa fa-trash'></i> </a>
+                                                        <?php
+                                                        }
+
+                                                        ?>
+
                                                     </td>
                                                 </tr>
                                                 <!-- EDIT NEW CATEGORY MODAL   -->
